refactor: Car 객체 수정
